requirements.txt
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
Django==4.2.11
django-cors-headers==4.4.0
django-redis==5.4.0
djangorestframework==3.15.1
gunicorn==22.0.0
idna==3.7
packaging==24.1
pillow==10.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
redis==5.0.4
requests==2.31.0
sqlparse==0.5.0
urllib3==2.2.2
